Added database support on the backend. Added ability to create and join groups. Client's are now recognized and remembered by IP (this can easily be changed to a username/pass combo) Added a new index.html where players can create and join their currently registered groups. (Just a concept, feel free to change this) Renamed the game page game.html. Players are routed to this page only when they join a group. (same as above) Minor changes to game.html to reflect a client's group info.
NOTE: In order for the database to work, you must have a MongoDB instance for the server to connect to. This is pretty easy. Just download Mongo from their website and run the database program. The server will automatically connect to the localhost and the rest is automated. You will need to tell Mongo where to store the database, however. On windows, this is done with the "--dbpath" parameter (it should be the same for OSX, but I'm not 100%). Let me know if you have any problems running this commit.